POC for 3.0 compatibility
* ENHANCEMENT: Added functionality to detect if level groups allow single or multiple levels and make the products available for checkout within a single cart.

For example if level 1 is in group 1 and group 1 only allows one level per group you may only add one to your cart per checkout. If you have a level that belongs to a group with multiple levels, you may add as many products as you'd like within this group.